+++ tests/testthat/test-spatial_nndm_cv.R
@@ -25,13 +25,12 @@ test_that("bad args", {
 
 test_that("can pass the dots to st_sample", {
   skip_if_not(sf::sf_use_s2())
-  expect_error(
+  expect_no_error(
     spatial_nndm_cv(
       Smithsonian_sf[1:15, ],
       Smithsonian_sf[16:20, ],
       type = "regular"
-    ),
-    NA
+    )
   )
 })
 
@@ -73,6 +72,31 @@ test_that("normal usage", {
     spatial_nndm_cv(Smithsonian_sf[1:15, ], Smithsonian_sf[16:20, ])
   )
 })
+
+test_that("can pass a single polygon to sample within", {
+  library(sf)
+  skip_if_not(sf::sf_use_s2())
+
+  example_poly <- sf::st_as_sfc(
+    list(
+      sf::st_point(c(-77.03, 40)),
+      sf::st_point(c(-76, 40.5)),
+      sf::st_point(c(-76.5, 39.5))
+    )
+  )
+  example_poly <- sf::st_set_crs(example_poly, sf::st_crs(Smithsonian_sf))
+  example_poly <- sf::st_union(example_poly)
+  example_poly <- sf::st_cast(example_poly, "POLYGON")
+
+  expect_snapshot(
+    spatial_nndm_cv(
+      Smithsonian_sf,
+      example_poly
+    )
+  )
+})
+
+
 
 test_that("printing", {
   skip_if_not(sf::sf_use_s2())